Separate fixtures into conftest.py
